Remove promisify module, wasnt promisifying nvim properly
